+++ pandas/tests/arithmetic/test_timedelta64.py
@@ -4,27 +4,11 @@ from datetime import (
     datetime,
     timedelta,
 )
-from itertools import (
-    chain,
-    combinations_with_replacement,
-    product,
-)
-from operator import attrgetter
-from typing import (
-    NamedTuple,
-    Type,
-    Union,
-)
 
-from hypothesis import given
-import hypothesis.strategies as st
 import numpy as np
 import pytest
 
-from pandas.errors import (
-    OutOfBoundsDatetime,
-    PerformanceWarning,
-)
+from pandas.errors import PerformanceWarning
 
 import pandas as pd
 from pandas import (
@@ -44,91 +28,11 @@ from pandas.core.api import (
     Int64Index,
     UInt64Index,
 )
-from pandas.core.arrays import (
-    DatetimeArray,
-    TimedeltaArray,
-)
 from pandas.tests.arithmetic.common import (
     assert_invalid_addsub_type,
     assert_invalid_comparison,
     get_upcast_box,
 )
-
-timedelta_types = (Timedelta, TimedeltaArray, TimedeltaIndex, Series, DataFrame)
-timestamp_types = (Timestamp, DatetimeArray, DatetimeIndex, Series, DataFrame)
-containers = slice(1, None)
-get_item_names = lambda t: "-".join(map(attrgetter("__name__"), t))
-
-
-class BinaryOpTypes(NamedTuple):
-    """
-    The expected operand and result types for a binary operation.
-    """
-
-    left: Type
-    right: Type
-    result: Type
-
-    def __str__(self) -> str:
-        return get_item_names(self)
-
-    def __repr__(self) -> str:
-        return f"BinaryOpTypes({self})"
-
-
-positive_tds = st.integers(min_value=1, max_value=Timedelta.max.value).map(Timedelta)
-
-xfail_no_overflow_check = pytest.mark.xfail(reason="No overflow check")
-
-
-@pytest.fixture(
-    name="add_sub_types",
-    scope="module",
-    params=tuple(combinations_with_replacement(timedelta_types, 2)),
-    ids=get_item_names,
-)
-def fixture_add_sub_types(request) -> BinaryOpTypes:
-    """
-    Expected types when adding, subtracting Timedeltas.
-    """
-    return_type = max(request.param, key=lambda t: timedelta_types.index(t))
-    return BinaryOpTypes(request.param[0], request.param[1], return_type)
-
-
-@pytest.fixture(
-    name="ts_add_sub_types",
-    scope="module",
-    params=tuple(product(timedelta_types, timestamp_types)),
-    ids=get_item_names,
-)
-def fixture_ts_add_sub_types(request) -> BinaryOpTypes:
-    """
-    Expected types when adding, subtracting Timedeltas and Timestamps.
-    """
-    type_hierarchy = {
-        name: i
-        for i, name in chain(enumerate(timedelta_types), enumerate(timestamp_types))
-    }
-    return_type = timestamp_types[max(type_hierarchy[t] for t in request.param)]
-
-    return BinaryOpTypes(request.param[0], request.param[1], return_type)
-
-
-def wrap_value(value: Union[Timestamp, Timedelta], type_):
-    """
-    Return value wrapped in a container of given type_, or as-is if type_ is a scalar.
-    """
-    if issubclass(type_, (Timedelta, Timestamp)):
-        return type_(value)
-
-    if issubclass(type_, pd.core.arrays.ExtensionArray):
-        box_cls = pd.array
-    elif issubclass(type_, pd.Index):
-        box_cls = pd.Index
-    else:
-        box_cls = type_
-
-    return type_(tm.box_expected([value], box_cls))
 
 
 def assert_dtype(obj, expected_dtype):
@@ -366,123 +270,6 @@ class TestTimedelta64ArrayComparisons:
 
 # ------------------------------------------------------------------
 # Timedelta64[ns] dtype Arithmetic Operations
-
-
-@given(positive_td=positive_tds)
-def test_add_raises_expected_error_if_result_would_overflow(
-    add_sub_types: BinaryOpTypes,
-    positive_td: Timedelta,
-):
-    left = wrap_value(Timedelta.max, add_sub_types.left)
-    right = wrap_value(positive_td, add_sub_types.right)
-
-    if add_sub_types.result is Timedelta:
-        msg = "|".join(
-            [
-                "int too big to convert",
-                "Python int too large to convert to C long",
-            ]
-        )
-    else:
-        msg = "Overflow in int64 addition"
-
-    with pytest.raises(OverflowError, match=msg):
-        left + right
-
-    with pytest.raises(OverflowError, match=msg):
-        right + left
-
-
-@xfail_no_overflow_check
-@given(positive_td=positive_tds)
-def test_sub_raises_expected_error_if_result_would_overflow(
-    add_sub_types: BinaryOpTypes,
-    positive_td: Timedelta,
-):
-    left = wrap_value(Timedelta.min, add_sub_types.left)
-    right = wrap_value(positive_td, add_sub_types.right)
-
-    msg = "Overflow in int64 addition"
-    with pytest.raises(OverflowError, match=msg):
-        left - right
-
-    with pytest.raises(OverflowError, match=msg):
-        (-1 * right) - abs(left)
-
-
-@given(td_value=positive_tds)
-def test_add_timestamp_raises_expected_error_if_result_would_overflow(
-    ts_add_sub_types: BinaryOpTypes,
-    td_value: Timedelta,
-):
-    left = wrap_value(td_value, ts_add_sub_types.left)
-    right = wrap_value(Timestamp.max, ts_add_sub_types.right)
-
-    ex = (OutOfBoundsDatetime, OverflowError)
-    msg = "|".join(["Out of bounds nanosecond timestamp", "Overflow in int64 addition"])
-
-    with pytest.raises(ex, match=msg):
-        left + right
-
-    with pytest.raises(ex, match=msg):
-        right + left
-
-
-@xfail_no_overflow_check
-@given(td_value=positive_tds)
-def test_sub_timestamp_raises_expected_error_if_result_would_overflow(
-    ts_add_sub_types: BinaryOpTypes,
-    td_value: Timedelta,
-):
-    right = wrap_value(td_value, ts_add_sub_types[0])
-    left = wrap_value(Timestamp.min, ts_add_sub_types[1])
-
-    ex = (OutOfBoundsDatetime, OverflowError)
-    msg = "|".join(["Out of bounds nanosecond timestamp", "Overflow in int64 addition"])
-
-    with pytest.raises(ex, match=msg):
-        left - right
-
-
-@given(value=st.floats().filter(lambda f: abs(f) > 1))
-def test_scalar_multiplication_raises_expected_error_if_result_would_overflow(
-    value: float,
-):
-    td = Timedelta.max
-
-    msg = "|".join(
-        [
-            "cannot convert float infinity to integer",
-            "Python int too large to convert to C long",
-            "int too big to convert",
-        ]
-    )
-    with pytest.raises(OverflowError, match=msg):
-        td * value
-
-    with pytest.raises(OverflowError, match=msg):
-        value * td
-
-
-@xfail_no_overflow_check
-@given(value=st.floats().filter(lambda f: abs(f) > 1))
-@pytest.mark.parametrize(
-    argnames="td_type",
-    argvalues=timedelta_types[containers],  # type: ignore[arg-type]
-    ids=attrgetter("__name__"),
-)
-def test_container_scalar_multiplication_raises_expected_error_if_result_would_overflow(
-    value: float,
-    td_type: Type,
-):
-    td = wrap_value(Timedelta.max, td_type)
-
-    msg = "Overflow in int64 addition"
-    with pytest.raises(OverflowError, match=msg):
-        td * value
-
-    with pytest.raises(OverflowError, match=msg):
-        value * td
 
 
 class TestTimedelta64ArithmeticUnsorted:
@@ -870,47 +657,6 @@ class TestTimedelta64ArithmeticUnsorted:
         )
         tm.assert_index_equal(result, exp)
         assert result.freq is None
-
-
-class TestAddSubNaTMasking:
-    # TODO: parametrize over boxes
-
-    @pytest.mark.parametrize("str_ts", ["1950-01-01", "1980-01-01"])
-    def test_tdarr_add_timestamp_nat_masking(self, box_with_array, str_ts):
-        # GH#17991 checking for overflow-masking with NaT
-        tdinat = pd.to_timedelta(["24658 days 11:15:00", "NaT"])
-        tdobj = tm.box_expected(tdinat, box_with_array)
-
-        ts = Timestamp(str_ts)
-        ts_variants = [
-            ts,
-            ts.to_pydatetime(),
-            ts.to_datetime64().astype("datetime64[ns]"),
-            ts.to_datetime64().astype("datetime64[D]"),
-        ]
-
-        for variant in ts_variants:
-            res = tdobj + variant
-            if box_with_array is DataFrame:
-                assert res.iloc[1, 1] is NaT
-            else:
-                assert res[1] is NaT
-
-    def test_tdi_add_overflow(self):
-        # These should not overflow!
-        exp = TimedeltaIndex([NaT])
-        result = pd.to_timedelta([NaT]) - Timedelta("1 days")
-        tm.assert_index_equal(result, exp)
-
-        exp = TimedeltaIndex(["4 days", NaT])
-        result = pd.to_timedelta(["5 days", NaT]) - Timedelta("1 days")
-        tm.assert_index_equal(result, exp)
-
-        exp = TimedeltaIndex([NaT, NaT, "5 hours"])
-        result = pd.to_timedelta([NaT, "5 days", "1 hours"]) + pd.to_timedelta(
-            ["7 seconds", NaT, "4 hours"]
-        )
-        tm.assert_index_equal(result, exp)
 
 
 class TestTimedeltaArraylikeAddSubOps:
